// __generated__/__types__.tsx
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
};

export type Analytics = {
  __typename?: 'Analytics';
  google_analytics?: Maybe<Scalars['String']>;
  umami_id?: Maybe<Scalars['String']>;
  simple_analytics?: Maybe<Scalars['String']>;
  plausible_data_domain?: Maybe<Scalars['String']>;
};

export type Author = {
  __typename?: 'Author';
  id: Scalars['Int'];
  email: Scalars['String'];
  username: Scalars['String'];
  name: Scalars['String'];
  social?: Maybe<Social>;
  bio?: Maybe<Scalars['String']>;
  occupation?: Maybe<Scalars['String']>;
  company_name?: Maybe<Scalars['String']>;
  role?: Maybe<Role>;
  permissions?: Maybe<Array<Permissions>>;
  avatar?: Maybe<Scalars['String']>;
  verified?: Maybe<Scalars['Boolean']>;
  accessToken?: Maybe<Scalars['String']>;
  analytics_id?: Maybe<Scalars['Int']>;
  analytics_uuid?: Maybe<Scalars['String']>;
};

export type AuthorNotFoundError = LetterpadError & {
  __typename?: 'AuthorNotFoundError';
  message: Scalars['String'];
};

export type AuthorResponse = {
  __typename?: 'AuthorResponse';
  ok: Scalars['Boolean'];
  errors?: Maybe<Array<Maybe<Error>>>;
  data?: Maybe<Author>;
};

export type CreateAuthorError = LetterpadError & {
  __typename?: 'CreateAuthorError';
  message: Scalars['String'];
};

export type CreateAuthorResponse = Author | CreateAuthorError;

export type CreateDomainResponse = Domain | DomainError;

export type CreatePostResponse = Post | PostError;


export type DeleteTagsResponse = DeleteTagsResult | TagsError;

export type DeleteTagsResult = {
  __typename?: 'DeleteTagsResult';
  ok: Scalars['Boolean'];
};

export type Domain = {
  __typename?: 'Domain';
  id: Scalars['Int'];
  name: Scalars['String'];
  mapped?: Maybe<Scalars['Boolean']>;
  ssl?: Maybe<Scalars['Boolean']>;
};

export type DomainError = LetterpadError & {
  __typename?: 'DomainError';
  message: Scalars['String'];
};

export type DomainNotFound = LetterpadError & {
  __typename?: 'DomainNotFound';
  message: Scalars['String'];
};

export type DomainResponse = Domain | DomainNotFound;

export type EditTaxResponse = {
  __typename?: 'EditTaxResponse';
  ok: Scalars['Boolean'];
};

export type Email = {
  __typename?: 'Email';
  template_id: Scalars['String'];
  body: Scalars['String'];
  subject: Scalars['String'];
};

export type EmailError = LetterpadError & {
  __typename?: 'EmailError';
  message: Scalars['String'];
};

export type EmailResponse = Email | EmailError;

export type Error = {
  __typename?: 'Error';
  path: Scalars['String'];
  message?: Maybe<Scalars['String']>;
};

export type ForgotPasswordResponse = {
  __typename?: 'ForgotPasswordResponse';
  ok: Scalars['Boolean'];
  message?: Maybe<Scalars['String']>;
};

export type Image = {
  __typename?: 'Image';
  src?: Maybe<Scalars['String']>;
  width?: Maybe<Scalars['Int']>;
  height?: Maybe<Scalars['Int']>;
};
// ...
export type LetterpadError = {
  message: Scalars['String'];
};

export type LoginData = {
  email: Scalars['String'];
  password: Scalars['String'];
};

export type LoginError = LetterpadError & {
  __typename?: 'LoginError';
  message: Scalars['String'];
};

export type LoginResponse = Author | LoginError;

export type MeResponse = Author | AuthorNotFoundError;
// ...
export type MediaNode = {
  __typename?: 'MediaNode';
  count: Scalars['Int'];
  rows: Array<Media>;
};
// ...
export enum NavigationType {
  Tag = 'tag',
  Page = 'page',
  Custom = 'custom'
}

export enum Permissions {
  ManageOwnPosts = 'MANAGE_OWN_POSTS',
  ReadOnlyPosts = 'READ_ONLY_POSTS',
  ManageAllPosts = 'MANAGE_ALL_POSTS',
  ManageUsers = 'MANAGE_USERS',
  ManageSettings = 'MANAGE_SETTINGS'
}

export type Post = {
  __typename?: 'Post';
  id: Scalars['Int'];
  title: Scalars['String'];
  html?: Maybe<Scalars['String']>;
  html_draft?: Maybe<Scalars['String']>;
  author?: Maybe<Author>;
  excerpt?: Maybe<Scalars['String']>;
  cover_image: Image;
  type: PostTypes;
  status: PostStatusOptions;
  featured: Scalars['Boolean'];
  slug?: Maybe<Scalars['String']>;
  createdAt: Scalars['Date'];
  publishedAt?: Maybe<Scalars['Date']>;
  scheduledAt?: Maybe<Scalars['Date']>;
  updatedAt: Scalars['Date'];
  reading_time?: Maybe<Scalars['String']>;
  tags?: Maybe<Array<Tags>>;
};

export type PostCountsByStatus = {
  __typename?: 'PostCountsByStatus';
  published: Scalars['Int'];
  drafts: Scalars['Int'];
};

export type PostError = LetterpadError & {
  __typename?: 'PostError';
  message: Scalars['String'];
};

export type PostFilters = {
  id?: Maybe<Scalars['Int']>;
  slug?: Maybe<Scalars['String']>;
  featured?: Maybe<Scalars['Boolean']>;
  previewHash?: Maybe<Scalars['String']>;
  status?: Maybe<PostStatusOptions>;
  type?: Maybe<PostTypes>;
};

export type PostResponse = Post | PostError;

export enum PostStatusOptions {
  Published = 'published',
  Draft = 'draft',
  Trashed = 'trashed'
}

export enum PostTypes {
  Post = 'post',
  Page = 'page'
}
// ...
export type PostsResponse = PostsNode | PostError;

export type Query = {
  __typename?: 'Query';
  domain: DomainResponse;
  email: EmailResponse;
  emails: Array<Maybe<Email>>;
  me?: Maybe<MeResponse>;
  media: MediaNode;
  post: PostResponse;
  posts: PostsResponse;
  settings: SettingResponse;
  stats?: Maybe<StatsResponse>;
  subscribers: SubscribersNode;
  tag: TagResponse;
  tags: TagsResponse;
};


export type QueryEmailArgs = {
  template_id?: Maybe<Scalars['String']>;
};


export type QueryMediaArgs = {
  filters?: Maybe<MediaFilters>;
};


export type QueryPostArgs = {
  filters?: Maybe<PostFilters>;
};


export type QueryPostsArgs = {
  filters?: Maybe<PostsFilters>;
};


export type QuerySubscribersArgs = {
  author_id?: Maybe<Scalars['Int']>;
};


export type QueryTagArgs = {
  slug: Scalars['String'];
};


export type QueryTagsArgs = {
  filters?: Maybe<TagsFilters>;
};
// ...
export type SettingResponse = Setting | SettingError;
// ...
export type StatsResponse = Stats | StatsError;

export type Subscribers = {
  __typename?: 'Subscribers';
  id: Scalars['Int'];
  email: Scalars['String'];
  author_id: Scalars['Int'];
  verified: Scalars['Boolean'];
  createdAt: Scalars['Date'];
};

export type SubscribersAddResult = {
  __typename?: 'SubscribersAddResult';
  ok: Scalars['Boolean'];
  message?: Maybe<Scalars['String']>;
};

export type SubscribersNode = {
  __typename?: 'SubscribersNode';
  count: Scalars['Int'];
  rows: Array<Subscribers>;
};

export type SubscribersUpdateResult = {
  __typename?: 'SubscribersUpdateResult';
  ok: Scalars['Boolean'];
  message?: Maybe<Scalars['String']>;
};

export type TagResponse = TagResultError | Tags;

export type TagResultError = LetterpadError & {
  __typename?: 'TagResultError';
  message: Scalars['String'];
};

export type Tags = {
  __typename?: 'Tags';
  name: Scalars['String'];
  slug: Scalars['String'];
  posts?: Maybe<PostsResponse>;
};

export type TagsError = LetterpadError & {
  __typename?: 'TagsError';
  message: Scalars['String'];
};

export type TagsFilters = {
  active?: Maybe<Scalars['Boolean']>;
  name?: Maybe<Scalars['String']>;
  suggest?: Maybe<Scalars['Boolean']>;
};

export type TagsInputType = {
  name: Scalars['String'];
  slug: Scalars['String'];
};

export type TagsNode = {
  __typename?: 'TagsNode';
  rows: Array<Tags>;
};

export type TagsResponse = TagsNode | TagsError;
// ...
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | ResolverWithResolve<TResult, TParent, TContext, TArgs>
  | StitchingResolver<TResult, TParent, TContext, TArgs>;
// ...
export type ResolversTypes = {
// ...
  Post: ResolverTypeWrapper<Post>;
// ...
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Analytics: Analytics;
  String: Scalars['String'];
  Author: Author;
  Int: Scalars['Int'];
  Boolean: Scalars['Boolean'];
  AuthorNotFoundError: AuthorNotFoundError;
  AuthorResponse: AuthorResponse;
  CreateAuthorError: CreateAuthorError;
  CreateAuthorResponse: ResolversParentTypes['Author'] | ResolversParentTypes['CreateAuthorError'];
  CreateDomainResponse: ResolversParentTypes['Domain'] | ResolversParentTypes['DomainError'];
  CreatePostResponse: ResolversParentTypes['Post'] | ResolversParentTypes['PostError'];
  Date: Scalars['Date'];
  DeleteTagsResponse: ResolversParentTypes['DeleteTagsResult'] | ResolversParentTypes['TagsError'];
  DeleteTagsResult: DeleteTagsResult;
  Domain: Domain;
  DomainError: DomainError;
  DomainNotFound: DomainNotFound;
  DomainResponse: ResolversParentTypes['Domain'] | ResolversParentTypes['DomainNotFound'];
  EditTaxResponse: EditTaxResponse;
  Email: Email;
  EmailError: EmailError;
  EmailResponse: ResolversParentTypes['Email'] | ResolversParentTypes['EmailError'];
  Error: Error;
  ForgotPasswordResponse: ForgotPasswordResponse;
  Image: Image;
  InputAnalytics: InputAnalytics;
  InputAuthor: InputAuthor;
  InputCreateAuthor: InputCreateAuthor;
  InputCreatePost: InputCreatePost;
  InputDomain: InputDomain;
  InputImage: InputImage;
  InputNavigation: InputNavigation;
  InputSocial: InputSocial;
  InputTags: InputTags;
  InputUpdateMedia: InputUpdateMedia;
  InputUpdatePost: InputUpdatePost;
  InputUpdateSubscriber: InputUpdateSubscriber;
  LetterpadError: ResolversParentTypes['AuthorNotFoundError'] | ResolversParentTypes['CreateAuthorError'] | ResolversParentTypes['DomainError'] | ResolversParentTypes['DomainNotFound'] | ResolversParentTypes['EmailError'] | ResolversParentTypes['LoginError'] | ResolversParentTypes['MediaError'] | ResolversParentTypes['PostError'] | ResolversParentTypes['SettingError'] | ResolversParentTypes['StatsError'] | ResolversParentTypes['TagResultError'] | ResolversParentTypes['TagsError'];
  LoginData: LoginData;
  LoginError: LoginError;
  LoginResponse: ResolversParentTypes['Author'] | ResolversParentTypes['LoginError'];
  MeResponse: ResolversParentTypes['Author'] | ResolversParentTypes['AuthorNotFoundError'];
  Media: Media;
  MediaDeleteResponse: ResolversParentTypes['MediaDeleteResult'] | ResolversParentTypes['MediaError'];
  MediaDeleteResult: MediaDeleteResult;
  MediaError: MediaError;
  MediaFilters: MediaFilters;
  MediaNode: MediaNode;
  MediaUpdateResponse: ResolversParentTypes['MediaUpdateResult'] | ResolversParentTypes['MediaError'];
  MediaUpdateResult: MediaUpdateResult;
  Mutation: {};
  Navigation: Navigation;
  Post: Post;
  PostCountsByStatus: PostCountsByStatus;
  PostError: PostError;
  PostFilters: PostFilters;
  PostResponse: ResolversParentTypes['Post'] | ResolversParentTypes['PostError'];
  PostsFilters: PostsFilters;
  PostsNode: PostsNode;
  PostsResponse: ResolversParentTypes['PostsNode'] | ResolversParentTypes['PostError'];
  Query: {};
  RemoveDomainResponse: RemoveDomainResponse;
  Response: Response;
  Setting: Setting;
  SettingError: SettingError;
  SettingInputType: SettingInputType;
  SettingResponse: ResolversParentTypes['Setting'] | ResolversParentTypes['SettingError'];
  Social: Social;
  Stats: Stats;
  StatsError: StatsError;
  StatsResponse: ResolversParentTypes['Stats'] | ResolversParentTypes['StatsError'];
  Subscribers: Subscribers;
  SubscribersAddResult: SubscribersAddResult;
  SubscribersNode: SubscribersNode;
  SubscribersUpdateResult: SubscribersUpdateResult;
  TagResponse: ResolversParentTypes['TagResultError'] | ResolversParentTypes['Tags'];
  TagResultError: TagResultError;
  Tags: Omit<Tags, 'posts'> & { posts?: Maybe<ResolversParentTypes['PostsResponse']> };
  TagsError: TagsError;
  TagsFilters: TagsFilters;
  TagsInputType: TagsInputType;
  TagsNode: TagsNode;
  TagsResponse: ResolversParentTypes['TagsNode'] | ResolversParentTypes['TagsError'];
  UpdateDomainResponse: UpdateDomainResponse;
  UpdatePostResponse: ResolversParentTypes['Post'] | ResolversParentTypes['PostError'];
  UpdateTagsResponse: ResolversParentTypes['EditTaxResponse'] | ResolversParentTypes['TagsError'];
};
// ...
export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  domain?: Resolver<ResolversTypes['DomainResponse'], ParentType, ContextType>;
  email?: Resolver<ResolversTypes['EmailResponse'], ParentType, ContextType, RequireFields<QueryEmailArgs, never>>;
  emails?: Resolver<Array<Maybe<ResolversTypes['Email']>>, ParentType, ContextType>;
  me?: Resolver<Maybe<ResolversTypes['MeResponse']>, ParentType, ContextType>;
  media?: Resolver<ResolversTypes['MediaNode'], ParentType, ContextType, RequireFields<QueryMediaArgs, never>>;
  post?: Resolver<ResolversTypes['PostResponse'], ParentType, ContextType, RequireFields<QueryPostArgs, never>>;
  posts?: Resolver<ResolversTypes['PostsResponse'], ParentType, ContextType, RequireFields<QueryPostsArgs, never>>;
  settings?: Resolver<ResolversTypes['SettingResponse'], ParentType, ContextType>;
  stats?: Resolver<Maybe<ResolversTypes['StatsResponse']>, ParentType, ContextType>;
  subscribers?: Resolver<ResolversTypes['SubscribersNode'], ParentType, ContextType, RequireFields<QuerySubscribersArgs, never>>;
  tag?: Resolver<ResolversTypes['TagResponse'], ParentType, ContextType, RequireFields<QueryTagArgs, 'slug'>>;
  tags?: Resolver<ResolversTypes['TagsResponse'], ParentType, ContextType, RequireFields<QueryTagsArgs, never>>;
};